backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
openai>=1.58.0
qdrant-client>=1.12.0
pydantic>=2.10.0
python-dotenv>=1.0.0
psycopg2-binary>=2.9.0
sqlalchemy>=2.0.36
alembic>=1.14.0
